修改分数为int

+++ microservice-functional/microservice-okr-core/src/main/java/com/aacoptics/okr/core/entity/po/KeyResultDetail.java
@@ -27,7 +27,7 @@ public class KeyResultDetail implements Serializable {
     private String keyResultName;
     private String remark;
     private Integer statusInfo;
-    private float score;
+    private Integer score;
     @TableLogic
     private String deleted = "N";
     private LocalDateTime createdTime;
